requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
python-gnupg==0.5.4
requests==2.32.3
smmap==5.0.2
tqdm==4.67.1
urllib3==2.4.0